Updated department, user, team role, invitation services

+++ src/main/java/com/theinnovationtrio/TeamFinderAPI/teamRole/TeamRoleService.java
@@ -6,10 +6,13 @@ import com.theinnovationtrio.TeamFinderAPI.user.User;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.stereotype.Service;
 
+import java.security.Principal;
 import java.util.List;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -19,13 +22,13 @@ public class TeamRoleService implements ITeamRoleService {
 
 
     @Override
-    public TeamRole createTeamRole(UUID userId, TeamRoleDto teamRoleDto) {
-        User adminUser = userService.getUserById(userId);
+    public TeamRole createTeamRole(Principal connectedUser, TeamRoleDto teamRoleDto) {
+        User adminUser = (User) ((UsernamePasswordAuthenticationToken) connectedUser).getPrincipal();
         boolean hasAdminRole = adminUser.getRoles().stream()
                 .anyMatch(role -> role.equals(Role.Organization_Admin));
         if (hasAdminRole) {
             TeamRole teamRole =
-                    new TeamRole(UUID.randomUUID(), teamRoleDto.getRoleInProject(), userId);
+                    new TeamRole(UUID.randomUUID(), teamRoleDto.getRoleInProject(), adminUser.getId());
             return teamRoleRepository.save(teamRole);
         } else {
             throw new AccessDeniedException("Unauthorized access!");
@@ -38,24 +41,38 @@ public class TeamRoleService implements ITeamRoleService {
     }
 
     @Override
+    public List<TeamRole> getAllSameOrgTeamRoles(Principal connectedUser) {
+        List<TeamRole> teamRoles = getAllTeamRoles();
+        User adminUser = (User) ((UsernamePasswordAuthenticationToken) connectedUser).getPrincipal();
+        boolean hasAdminRole = adminUser.getRoles().stream()
+                .anyMatch(role -> role.equals(Role.Organization_Admin));
+        if(hasAdminRole){
+            return teamRoles.stream()
+                    .filter(teamRole -> userService.getUserById(teamRole.getCreatedBy()).getOrganizationId()
+                            .equals(adminUser.getOrganizationId()))
+                    .collect(Collectors.toList());
+        } else {
+            throw new AccessDeniedException("Unauthorized access!");
+        }
+    }
+
+    @Override
     public TeamRole getTeamRoleById(UUID teamRoleId) {
         return teamRoleRepository.findById(teamRoleId).orElseThrow(() -> new EntityNotFoundException("Team Role not found!"));
     }
 
     @Override
-    public TeamRole updateTeamRole(UUID userId, UUID teamRoleId, TeamRoleDto teamRoleDto) {
-        User adminUser = userService.getUserById(userId);
+    public TeamRole updateTeamRole(Principal connectedUser, UUID teamRoleId, TeamRoleDto teamRoleDto) {
+        User adminUser = (User) ((UsernamePasswordAuthenticationToken) connectedUser).getPrincipal();
         boolean hasAdminRole = adminUser.getRoles().stream()
                 .anyMatch(role -> role.equals(Role.Organization_Admin));
-        if(hasAdminRole){
-            try{
-                TeamRole teamRole = getTeamRoleById(teamRoleId);
-                teamRole.setCreatedBy(userId);
+        TeamRole teamRole = getTeamRoleById(teamRoleId);
+        boolean hasSameOrganization = userService.getUserById(teamRole.getCreatedBy()).getOrganizationId()
+                .equals(adminUser.getOrganizationId());
+        if(hasAdminRole && hasSameOrganization){
+                teamRole.setCreatedBy(adminUser.getId());
                 teamRole.setRoleInProject(teamRoleDto.getRoleInProject());
                 return teamRoleRepository.save(teamRole);
-            } catch (EntityNotFoundException ex) {
-                throw new EntityNotFoundException("Team role does not exist.");
-            }
         } else {
             throw new AccessDeniedException("Unauthorized access!");
         }
@@ -63,17 +80,15 @@ public class TeamRoleService implements ITeamRoleService {
 
 
     @Override
-    public void deleteTeamRole(UUID userId, UUID teamRoleId) {
-        User adminUser = userService.getUserById(userId);
+    public void deleteTeamRole(Principal connectedUser, UUID teamRoleId) {
+        User adminUser = (User) ((UsernamePasswordAuthenticationToken) connectedUser).getPrincipal();
         boolean hasAdminRole = adminUser.getRoles().stream()
                 .anyMatch(role -> role.equals(Role.Organization_Admin));
-        if(hasAdminRole){
-            try{
-                getTeamRoleById(teamRoleId);
+        TeamRole teamRole = getTeamRoleById(teamRoleId);
+        boolean hasSameOrganization = userService.getUserById(teamRole.getCreatedBy()).getOrganizationId()
+                .equals(adminUser.getOrganizationId());
+        if(hasAdminRole && hasSameOrganization){
                 teamRoleRepository.deleteById(teamRoleId);
-            } catch (EntityNotFoundException ex) {
-                throw new EntityNotFoundException("Team role does not exist.");
-            }
         } else {
             throw new AccessDeniedException("Unauthorized access!");
         }
